+++ gems/kettle-jem/lib/kettle/jem/tasks/install_task.rb
@@ -16,7 +16,7 @@ module Kettle
       module InstallTask
         module_function
 
-        CURATED_BINSTUB_GEMS = %w[appraisal2 rake rbs rspec-core yard kettle-dev kettle-test kettle-soup-cover kettle-gha-pins stone_checksums].freeze
+        CURATED_BINSTUB_GEMS = %w[appraisal2 rake rbs rspec-core yard rubocop-gradual kettle-dev kettle-test kettle-soup-cover kettle-gha-pins stone_checksums].freeze
         CURATED_BINSTUB_DISCRETE_EXECUTABLES = %w[
           kettle-drift
           yard